dao: use long keys for text messages
Text message IDs, coming from the daemon, are 64 bits.
Store them as long in the database. This change is backward compatible.

+++ ring-android/androidtv/src/main/java/cx/ring/services/HistoryServiceImpl.java
@@ -75,7 +75,7 @@ public class HistoryServiceImpl extends HistoryService {
     }
 
     @Override
-    protected Dao<HistoryText, Integer> getTextHistoryDao() {
+    protected Dao<HistoryText, Long> getTextHistoryDao() {
         try {
             return getHelper().getTextHistoryDao();
         } catch (SQLException e) {
